requirements.txt
# GUI Dependencies
sv-ttk==2.5.5
pillow==10.2.0
pystray==0.19.5

# Core Functionality
keyboard==0.13.5
pyperclip==1.8.2
pywinstyles==1.1.0
elevate==0.1.3

# Type Hints Support
typing-extensions==4.9.0

# Optional but recommended for Windows
pywin32==306; platform_system == "Windows"